Added project Description
# Team Rocket

##Project Description:

As the video game industry grows, developers seek to provide more and more immersive experiences to their players. Many companies, such as Nintendo, have attempted to bring gameplay experiences into the physical realm through devices such as the Rumble Pak and Nintendo Labo. Few developers, however, have tried to bring the experience of flight to their players. This project provides haptic feedback for a flight simulation game. While strapped to a wooden frame, the player uses a Wii balance board to control the motion of a hang glider in a Unity game. As the hang glider turns to the left or right, stepper motors attached to string and Velcro straps pull on the player’s arms to simulate wind resistance. 

## Materials used:
* Unity 
* [SerialCommUnity](https://github.com/dwilches/SerialCommUnity)
* Arduino Duemilanove
* Adafruit Motor Shield
* [Wantai Stepper Motor (2)](https://www.sparkfun.com/products/13656)
* Wii Balance Board
* [Wii Balance Walker Driver](http://www.greycube.com/site/download.php?view.68)


##Team members:
* Kyle Gronberg (Initial Project Build, Unity Code Tweaks) 
* Cade Haley (Final Build Tweaks, Arduino Code) 
* Madison Razook (Final Build Tweaks, Unity Code)